🚚 [RUMF-1097][logs] move LogsInitConfiguration and HybridConfiguration
